==> songwriter/src/01-get_data.ipynb <==
import tqdm
import time
import multiprocessing
import requests
import pandas as pd

API_KEY = "660a4395f992ff67786584e238f501aa"
ARTIST_URL = "https://www.vagalume.com.br/{artist}/index.js"
HEADERS = {'User-Agent': 'Mozilla/5.0'}

---

Example
---

url = ARTIST_URL.format(artist="roberto-carlos")
resp = requests.get(url=url, headers=HEADERS)

len(resp.json()['artist']['toplyrics']['item'])

resp = requests.get(url="https://api.vagalume.com.br/search.php?art=roberto-carlos&mus=nossa-senhora-2")

resp.json()

---

# Download Top 100 Artists

period = 202003
limit = 100
RANK_URL = f"https://api.vagalume.com.br/rank.php?apikey=660a4395f992ff67786584e238f501aa&type=art&period=month&periodVal={period}&limit={limit}"

top_100_artists = requests.get(url=RANK_URL, headers=HEADERS)

artists_df = pd.DataFrame.from_dict(top_100_artists.json()['art']['month']['all'])

artists_df["selecionado"] = 1

artists_df.to_csv("../data/artistas.csv", index=False, sep=";")

---

# List of top artists

# read updated artistas with selected ones
artists_df = pd.read_csv("../data/artistas.csv", sep=";")

artists_df.head()

artists_df = artists_df.query("selecionado == 1")

artist_url_list = artists_df["url"].values

artist_url_list[:3]

def get_artist_songs(artist_url: str) -> pd.DataFrame:
    url = artist_url + "index.js"
    resp = requests.get(url=url, headers=HEADERS)
    
    songs_df = pd.DataFrame.from_dict(resp.json()['artist']['toplyrics']['item'])
    songs_df["url_artista"] = artist_url
    
    return songs_df

with multiprocessing.Pool(processes=8) as pool:
    results = pool.map(get_artist_songs, artist_url_list)

songs_df = pd.concat(results)

songs_df = songs_df.merge(artists_df, left_on="url_artista", right_on="url", suffixes=["_musica", "_artista"])

songs_df.to_csv("../data/musicas.csv", sep=";", index=False)

---

# Get Lyrics

MUSIC_URL = "https://api.vagalume.com.br/search.php?musid={musid}"

songs_df = pd.read_csv("../data/musicas.csv", sep=";")

songs_df.head()

songs_df.shape

musid_list = songs_df["id_musica"].values

musid_list[:4]

def get_lyrics(musid: str) -> pd.DataFrame:
    try:
        url = MUSIC_URL.format(musid=musid)
        resp = requests.get(url=url, headers=HEADERS)

        lyrics_df = pd.DataFrame.from_dict(resp.json()['mus'])    
        
        # sleep so we don't get our IP blocked
        time.sleep(5)
        
        return lyrics_df
    except Exception as e:
        print(e, musid)
        return pd.DataFrame()

lyrics_results = []

# Wont be able to use parallelism because the site is blocking our requests
for mid in tqdm.tqdm(musid_list):
    lyrics_results.append(get_lyrics(mid))

lyrics_df = pd.concat(lyrics_results)

lyrics_df = lyrics_df.merge(songs_df, left_on="id", right_on="id_musica", suffixes=["_letra", "_musica"])

lyrics_df.to_csv("../data/letras.csv", sep=";", index=False)

lyrics_df.head()

---

lyrics_df = pd.read_csv("../data/letras.csv", sep=";")

final_df = lyrics_df[["name_musica", "name_letra", "text", "lang"]]
final_df.columns = ["artista", "musica", "letra", "lang"]
final_df.head()

final_df.to_csv("../data/data_raw.csv", sep=";", index=False)
