==> songwriter/src/02-build_dataset.ipynb <==
import pandas as pd

df = pd.read_csv("../data/data_raw.csv", sep=";")

df.shape

df.head()

df.value_counts("artista")

df["artista"].sort_values().unique()

df["lang"].value_counts()

---

Create a unique artist tag and add is to each new paragraph on the lyrics. 
This should guide the song writing by the model.

def get_artist_tag(artist_name):
    return "x_" + ''.join([w[:2] + w[-2:] for w in artist_name.lower().split(" ")]) + "_x"

# df["artist_tag"] = df["artista"].apply(lambda x: "xx_" + x.lower().replace(" ", "_") + "_xx")
df["artist_tag"] = df["artista"].apply(
    lambda x: get_artist_tag(x)
)

# there is one conflict
df["artist_tag"].nunique()

df["artist_lyrics"] = df.apply(
    lambda x: [x["artist_tag"] + " " + phrase for phrase in x["letra"].split("\n") if len(phrase) > 3], 
    axis=1
)

df["clean_lyrics"] = df["letra"].apply(lambda x: [phrase for phrase in x.split("\n") if len(phrase) > 3])

df.sample(5)

lyrics_phrase = pd.Series([l for lyrics in df.query("lang == 2")["artist_lyrics"] for l in lyrics], name="text")
clean_lyrics_phrase = pd.Series([l for lyrics in df.query("lang == 2")["clean_lyrics"] for l in lyrics], name="text")

df.to_csv("../data/lyrics_processed.csv", sep=";", index=False)

lyrics_phrase.to_csv("../data/lyrics_phrase_eng.csv", sep=";", index=False)
clean_lyrics_phrase.to_csv("../data/clean_lyrics_phrase_eng.csv", sep=";", index=False)

lyrics_phrase = pd.Series([l for lyrics in df.query("lang == 1")["artist_lyrics"] for l in lyrics], name="text")
clean_lyrics_phrase = pd.Series([l for lyrics in df.query("lang == 1")["clean_lyrics"] for l in lyrics], name="text")

lyrics_phrase.to_csv("../data/lyrics_phrase_pt.csv", sep=";", index=False)
clean_lyrics_phrase.to_csv("../data/clean_lyrics_phrase_pt.csv", sep=";", index=False)
