==> songwriter/src/02b-build_dataset_dl.ipynb <==
import string

import numpy as np
import pandas as pd

from nltk import word_tokenize

df = pd.read_csv("../data/clean_lyrics_phrase_eng.csv", sep=";")

string.punctuation

processed_text = df["text"].apply(
    lambda x: "".join([f" {c} " if c in string.punctuation else c for c in x.lower().strip()])
)

n = len(processed_text)

train_size = int(n*0.6)
valid_size = int(n*0.15)

np.savetxt("../data/ae/train.txt", processed_text.iloc[:train_size].values, fmt="%s")
np.savetxt("../data/ae/valid.txt", processed_text.iloc[train_size:train_size+valid_size].values, fmt="%s")
np.savetxt("../data/ae/test.txt", processed_text.iloc[train_size+valid_size:].values, fmt="%s")
