==> songwriter/src/04-autoencoder.ipynb <==
import pickle
import string

import numpy as np
import pandas as pd
from sklearn.manifold import TSNE
from sklearn.metrics.pairwise import cosine_distances

from text_autoencoders import TrainedAutoEncoder

ae = TrainedAutoEncoder("../src/text_autoencoders/checkpoints/lyrics/daae/")

---

TODO:
- [x] relate phrases with lyrics and artitst (may have to rebuild the set)
- [x] encode phrases in lyrics
- [x] get the mean vector of each lyric
- [ ] get the mean vector of each artist
- [x] reduce dim for lyrics and artists
- [x] create map of lyric -> 2dim
- [ ] create map of artist -> 2dim
- [x] plot lyrics space
- [ ] plot artists space

df = pd.read_csv("../data/data_raw.csv", sep=";")
df["clean_lyrics"] = df["letra"].apply(lambda x: [phrase for phrase in x.split("\n") if len(phrase) > 3])
df = df.reset_index()

lyrics_df = pd.DataFrame([(i, l) for i, lyrics in df.query("lang == 2")[["index", "clean_lyrics"]].values for l in lyrics], columns=["song_idx", "sentence"])
lyrics_df = lyrics_df.merge(df, left_on="song_idx", right_on="index").drop(columns=["clean_lyrics", "index"])

lyrics_df.head()

def process_text(text):
    return "".join([f" {c} " if c in string.punctuation else c for c in text.lower().strip()])

---

lyrics_df["sentence"].apply(process_text).str.split(" ").values

sentence_encodings = ae.encode(lyrics_df["sentence"].apply(process_text).str.split(" ").values)

sentence_encodings.shape

song_idxs = lyrics_df.reset_index().groupby("song_idx")["index"].apply(list)

song_encodings_map = {}

for i, idxs in song_idxs.iteritems():
    song_encodings_map[i] = sentence_encodings[idxs].mean(axis=0)

with open("../data/processed/sentence_encodings.pkl", "wb") as outfile:
    pickle.dump(sentence_encodings, outfile)

with open("../data/processed/song_encodings.pkl", "wb") as outfile:
    pickle.dump(song_encodings_map, outfile)

lyrics_df.to_csv("../data/processed/lyrics.csv", sep=";", index=False)

len(song_encodings_map)

---

lyrics_df.sample()

distances = cosine_distances(sentence_encodings[81990].reshape(1, -1), sentence_encodings)

lyrics_df.iloc[distances.argsort()[0]].head()

Distances seens to work

random_idx = np.random.randint(0, lyrics_df.shape[0])
distances = cosine_distances(sentence_encodings[random_idx].reshape(1, -1), sentence_encodings)
lyrics_df.iloc[distances.argsort()[0]].head()
