==> songwriter/src/06-song_interpolation.ipynb <==
import pickle

import numpy as np
import pandas as pd

from text_autoencoders import TrainedAutoEncoder

ae = TrainedAutoEncoder("../src/text_autoencoders/checkpoints/lyrics/daae/")

lyrics_df = pd.read_csv("../data/processed/lyrics.csv", sep=";")
lyrics_df.sample(3)

with open("../data/processed/sentence_encodings.pkl", "rb") as infile:
    sentence_encodings = pickle.load(infile)

sentence_encodings.shape

---

lyrics_df.sample()

song_1 = 1388
song_1_lyrics = lyrics_df.query("song_idx == @song_1")
song_1_lyrics.head(3)

song_2 = 352
song_2_lyrics = lyrics_df.query("song_idx == @song_2")
song_2_lyrics.head(3)

---

encoding_s1 = sentence_encodings[song_1_lyrics.index.tolist()]
encoding_s2 = sentence_encodings[song_2_lyrics.index.tolist()]

n_phrases = min(len(encoding_s1), len(encoding_s2))
n_phrases

def interpolate(e1, e2, n, w1=0.5, w2=0.5, shuffle=False):
    if shuffle:
        np.random.shuffle(e1)
        np.random.shuffle(e2)
        
    return (
        (w1 * e1[:n,:] + w2 * e2[:n,:]) / (w1+w2)
    )

interpolated_enc = interpolate(encoding_s1, encoding_s2, n_phrases, 0.4, 0.6, True)

interpolate_dec = ae.decode(interpolated_enc)

new_song = '\n'.join([' '.join(p).replace('<unk>', '').strip() for p in interpolate_dec])
print(new_song)

new_song = '\n'.join([' '.join(p).replace('<unk>', '').strip() for p in interpolate_dec])
print(new_song)
